+++ app/config.py
@@ -1,0 +1,5 @@
+# Config
+from os.path import abspath, dirname
+basedir = abspath(dirname(__file__))
+CRF_ENABLED = True
+SECRET_KEY = 'EXAMPLE_SECRET_KEY_HERE'
