+++ app/imnotcrazy.py
@@ -1,0 +1,6 @@
+from flask import Flask
+app = Flask(__name__)
+
+@app.route('/')
+def hello():
+    return "Hello world!"
